separate better the utils between forte and the interpreter
